# db: optimize levelIter for non-matching bloom filter